feat: split socket from app
