+++ src/index.ts
@@ -1,24 +1,14 @@
-import express from "express";
-import app, {server} from "./app.js";
-import {connect} from "mongoose";
-import bodyParser from "body-parser";
-import cors from "cors";
+import { connect } from "mongoose";
 import "./config";
 import "./sockets/discord.js";
-import {registerMessageIndexes} from "./sockets/discord.js";
-import {GuildDatabaseCreator} from "./services/guild/guildDatabaseCreator.js";
-import {errorHandler} from "./middleware/errorHandler.middleware.js";
-import {mapEndpoints} from "./endpoints.js";
-
-app.use(express.json());
-app.use(cors());
-
-app.use(bodyParser.urlencoded({extended: true}));
+import { registerMessageIndexes } from "./sockets/discord.js";
+import { GuildDatabaseCreator } from "./services/guild/guildDatabaseCreator.js";
+import { server } from "./socket.js";
 
 // Connect to database
 try {
     const dbUrl = process.env.DB_URL;
-    connect(dbUrl, {retryWrites: true, writeConcern: {w: "majority"}}).then(() => {
+    connect(dbUrl, { retryWrites: true, writeConcern: { w: "majority" } }).then(() => {
         const databaseCreator = new GuildDatabaseCreator();
         databaseCreator.registerDatabases().then(() => {
             registerMessageIndexes();
@@ -32,6 +22,3 @@ try {
 } catch (error) {
     console.error("Failed to connect to database:", error);
 }
-
-mapEndpoints(app);
-app.use(errorHandler);
